LayoutTests/platform/gtk/accessibility/svg-group-element-with-title-expected.txt

This tests SVG group elements are accessible and that the svg:title element is returned properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Verify that the SVG group is returned as an accessible element.
PASS group1.role is 'AXRole: AXGroup'
PASS group1.title is 'AXTitle: SVG TITLE 1'

Verify that you can hit-test to the SVG group.
PASS hitTestElement.isEqual(group1) is true
PASS successfullyParsed is true

TEST COMPLETE
